+++ module/actor/actor-sheet.mjs
@@ -917,7 +917,7 @@ export class HeroSystemActorSheet extends FoundryVttActorSheet {
         speaker.alias = actor.name;
 
         const chatData = {
-            style: CONST.CHAT_MESSAGE_STYLES.OOC,
+            style: CONST.CHAT_MESSAGE_STYLES.IC, //CONST.CHAT_MESSAGE_STYLES.OOC
             rolls: heroRoller.rawRolls(),
             author: game.user._id,
             content: cardHtml,
@@ -956,7 +956,7 @@ export class HeroSystemActorSheet extends FoundryVttActorSheet {
         speaker.alias = this.actor.name;
 
         const chatData = {
-            style: CONST.CHAT_MESSAGE_STYLES.OOC,
+            style: CONST.CHAT_MESSAGE_STYLES.IC, //CONST.CHAT_MESSAGE_STYLES.OOC
             rolls: characteristicRoller.rawRolls(),
             author: game.user._id,
             content: cardHtml,
@@ -1038,7 +1038,7 @@ export class HeroSystemActorSheet extends FoundryVttActorSheet {
         speaker.alias = this.actor.name;
 
         const chatData = {
-            style: CONST.CHAT_MESSAGE_STYLES.OOC,
+            style: CONST.CHAT_MESSAGE_STYLES.IC, //CONST.CHAT_MESSAGE_STYLES.OOC
             rolls: characteristicRoller.rawRolls(),
             author: game.user._id,
             content: cardHtml,