+++ module/utility/presence-attack.mjs
@@ -145,7 +145,7 @@ async function presenceAttackRoll(actor, html) {
     const cardHtml = await foundryVttRenderTemplate(template, cardData);
 
     const chatData = {
-        style: CONST.CHAT_MESSAGE_STYLES.OOC,
+        style: CONST.CHAT_MESSAGE_STYLES.IC, //CONST.CHAT_MESSAGE_STYLES.OOC
         rolls: preAttackRoller.rawRolls(),
         author: game.user._id,
         content: cardHtml,